✨ [Feat] 운동 ADMIN, USER 로직 분리 [WT]: back-end [RV]: motojate - 운동 ADMIN, USER 로직 분리
